population per tick implement

+++ DTO/src/impl/EntitiesAmountDTO.java
@@ -3,13 +3,13 @@ package impl;
 import java.util.Map;
 
 public class EntitiesAmountDTO {
-    private final Map<String, Integer> entityNameToAmount;
+    private final Map<String, Long> entityNameToAmount;
 
-    public EntitiesAmountDTO(Map<String, Integer> entityNameToAmount) {
+    public EntitiesAmountDTO(Map<String, Long> entityNameToAmount) {
         this.entityNameToAmount = entityNameToAmount;
     }
 
-    public Map<String, Integer> getEntityToPopulationMap() {
+    public Map<String, Long> getEntityToPopulationMap() {
         return entityNameToAmount;
     }
 }
